replace st_centroid with st_point_on_surface
The problem causing the NAs for some SA1s, even after adding centroids was that some polygons had some odd shape that meant that the centroid of the largest polygon was not necessarily within the polygon. This meant that the spatial join to interpolated values didn't get a match and we get NAs for the aggregate measures.